Refactor file structure and comments
