+++ app/bot/views/link_view.py
@@ -1,3 +1,5 @@
+"""Views for file and folder download links."""
+
 from textwrap import dedent
 
 from telethon import Button
@@ -9,9 +11,9 @@ from app.utils.language import Translator
 def render_file_link_message(file_result, translator: Translator) -> ViewResponse:
     """Render the file download link message."""
     message = dedent(f"""
-        <b>{translator.get('fileEmoji')} {file_result.name}</b>
+        <b>{translator.get("fileEmoji")} {file_result.name}</b>
 
-        <b>{translator.get('downloadLinkLabel')}</b>:
+        <b>{translator.get("downloadLinkLabel")}</b>:
         <code>{file_result.url}</code>
     """)
     buttons = [[Button.url(translator.get("downloadFileBtn"), file_result.url)]]
@@ -21,9 +23,9 @@ def render_file_link_message(file_result, translator: Translator) -> ViewRespons
 def render_folder_link_message(archive_url: str, translator: Translator) -> ViewResponse:
     """Render the folder download link message."""
     message = dedent(f"""
-        <b>{translator.get('folderEmoji')} {translator.get('folderDownloadLinkLabel')}</b>
+        <b>{translator.get("folderEmoji")} {translator.get("folderDownloadLinkLabel")}</b>
 
-        <b>{translator.get('downloadLinkLabel')}</b>:
+        <b>{translator.get("downloadLinkLabel")}</b>:
         <code>{archive_url}</code>
     """)
     buttons = [[Button.url(translator.get("downloadFolderBtn"), archive_url)]]
